+++ examples/switch.py
@@ -5,8 +5,8 @@ from rich import print
 
 from powercli import Command
 from powercli.methods import Switch
+from powercli.static import Static
 from powercli.typedefs import Context
-from powercli.utils import static
 
 logger.enable("powercli.command")
 logger.add(sys.stdout, level="TRACE")
@@ -20,12 +20,12 @@ cmd: Command[str, None] = Command()
 cmd.flag(
     identifier="f",
     short="f",
-    method=Switch(on_presence=static("A"), on_absence=static("B")),
+    method=Switch(on_presence=Static("A"), on_absence=Static("B")),
 )
 cmd.flag(
     identifier="no-color",
     long="no-color",
-    method=Switch(on_presence=disable_color, on_absence=static(None)),
+    method=Switch(on_presence=disable_color, on_absence=Static(None)),
 )
 
 if __name__ == "__main__":