+++ powercli/__init__.py
@@ -5,8 +5,9 @@
 
 from __future__ import annotations
 
-__all__ = ["Argument", "Flag", "Positional", "Category", "Command"]
+__all__ = ["Argument", "Flag", "Positional", "Category", "Command", "Static"]
 
 from .args import Argument, Flag, Positional
 from .category import Category
 from .command import Command
+from .static import Static