+++ powercli/static.py
@@ -1,0 +1,34 @@
+from __future__ import annotations
+
+from attrs import frozen
+from typing import Any
+from .typedefs import Context
+
+@frozen
+class Static[T]:
+    """
+    A wrapper around a value that when called with a single argument (the
+    context) simply returns the wrapped value.
+
+
+    This is useful when used for attributes where the actual context is
+    irrelevant. On top of that external code such as the help message generator
+    can improve the output if a static object is used.
+
+    # Examples
+
+    ```python
+    from powercli import Flag
+
+    Flag(
+        # ...
+        required=static(True)
+    )
+    ```
+    """
+
+    value: T
+    """The wrapped value that is returned when this object is called."""
+
+    def __call__(self, context: Context[Any, Any], *args: Any, **kwargs: Any) -> T:
+        return self.value