+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Changed
 
+- Deprecate `powercli.utils.static` in favor of `powercli.static.Static`.
 - Better support for inspecting subcommand arguments (via
   `powercli.parser.ParsedCommand.subcommand()`).
 - Usage section in help now includes parent commands names.